+++ examples/encode_decode_simple/encode_decode_simple.c
@@ -16,8 +16,10 @@
 /// Simple example showing how to encode and decode a block
 /// of memory.
 
-void trace_callback(const char* zone, const char* data)
+void trace_callback(const char* zone, const char* data, void* context)
 {
+    (void) context;
+
     if (strcmp(zone, "decoder_state") == 0)
     {
         printf("%s:\n", zone);
@@ -108,7 +110,7 @@ int main()
     // Install a custom trace function for the decoder (if tracing is enabled)
     if (kodo_has_set_trace_callback(decoder))
     {
-        kodo_set_trace_callback(decoder, trace_callback);
+        kodo_set_trace_callback(decoder, trace_callback, NULL);
     }
 
     while (!kodo_is_complete(decoder))