+++ examples/sliding_window/sliding_window.c
@@ -20,8 +20,9 @@
 /// such that symbols that have already been received at the decoder
 /// are not included in the encoding again (saving computations).
 
-void trace_callback(const char* zone, const char* data)
+void trace_callback(const char* zone, const char* data, void* context)
 {
+    (void) context;
     if (strcmp(zone, "decoder_state") == 0 ||
         strcmp(zone, "input_symbol_coefficients") == 0)
     {
@@ -90,7 +91,7 @@ int main()
     // Install a custom trace function for the decoder (if tracing is enabled)
     if (kodo_has_set_trace_callback(decoder))
     {
-        kodo_set_trace_callback(decoder, trace_callback);
+        kodo_set_trace_callback(decoder, trace_callback, NULL);
     }
 
     while (!kodo_is_complete(decoder))